Feature admin
- Admin page divided into Admin, Users and Metrics for easier navigation and code sustainability
