+++ src/service/userService.js
@@ -1,0 +1,6 @@
+import axios from './axios';
+
+export const fetchUsers = async () => {
+    const response = await axios.get('/users');
+    return response.data;
+};
